dockerfile

+++ src/main/java/com/liwell/cinema/domain/po/CollectDetail.java
@@ -80,7 +80,7 @@ public class CollectDetail {
     private String vod_sub;
     private String vod_tag;
     private String vod_time; // 更新时间
-    private Integer vod_time_add; // 更新时间戳
+    private Long vod_time_add; // 上映时间戳
     private Integer vod_time_hits;
     private Integer vod_time_make;
     private Integer vod_total; // 总集数
